+++ resources/views/beatmap_discussions/_item.blade.php
@@ -17,21 +17,31 @@
 --}}
 <ul>
     <li>
-        <a href="{{ route('beatmapsets.discussion', $post->beatmapset) }}#/{{ $post->beatmapDiscussion->getKey() }}">
-            {{ trans('admin.beatmap_discussions.item.permalink') }}
+        <a href="{{ route('beatmapsets.discussion', $discussion->beatmapset) }}#/{{ $discussion->getKey() }}">
+            {{ trans('beatmap_discussions.item.permalink') }}
         </a>
     </li>
 
     <li>
-        {!! link_to_user($post->user) !!}
+        {!! link_to_user($discussion->user) !!}
     </li>
 
     <li>
-        {{ trans('admin.beatmap_discussion_posts.item.content') }}:
-        @if ($post->system)
-            <pre>{{ json_encode($post->message) }}</pre>
-        @else
-            <pre>{{ $post->message }}</pre>
-        @endif
+        {{ trans('beatmap_discussions.item.message_type') }}: {{ $discussion->message_type ?? 'invalid' }}
+    </li>
+
+    <li>
+        {{ trans('beatmap_discussions.item.created_at') }}: {{ $discussion->created_at }}
+    </li>
+
+    @if ($discussion->deleted_at !== null)
+        <li>
+            {{ trans('beatmap_discussions.item.deleted_at') }}: {{ $discussion->deleted_at }}
+        </li>
+    @endif
+
+    <li>
+        {{ trans('beatmap_discussion_posts.item.content') }}:
+        <pre>{{ $discussion->startingPost->message }}</pre>
     </li>
 </ul>
